Refactor(Constructor) : 프로젝트 구조를 실무 구조와 유사하게 변경, 파일 위치 재배치
