--- src/main/java/com/multi/travel/api/repository/TourSpotApiRepository.java
@@ -1,15 +1,0 @@
-package com.multi.travel.api.repository;
-
-/*
- * Please explain the class!!!
- *
- * @filename    : TourSpotApiRepository
- * @author      : Choi MinHyeok
- * @since       : 25. 11. 8. 토요일
- */
-
-import com.multi.travel.tourspot.entity.TourSpot;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TourSpotApiRepository extends JpaRepository<TourSpot, Long> {
-}
