+++ src/main/java/com/multi/travel/auth/controller/AuthController.java
@@ -1,0 +1,13 @@
+package com.multi.travel.auth.controller;
+
+/*
+ * Please explain the class!!!
+ *
+ * @filename    : AuthController
+ * @author      : Choi MinHyeok
+ * @since       : 25. 12. 1. 월요일
+ */
+
+
+public class AuthController {
+}
