+++ src/main/java/com/multi/travel/auth/jwt/JwtProvider.java
@@ -1,0 +1,26 @@
+package com.multi.travel.auth.jwt;
+
+import io.jsonwebtoken.security.Keys;
+import lombok.Getter;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.util.Base64;
+
+@Component
+public class JwtProvider {
+
+    @Getter
+    @Value("${jwt.issuer}")
+    private String issuer;
+    @Value("${jwt.secret}")
+    private String secretKey;
+
+    public SecretKey getSecretKey() {
+        // Secret Key Base64 디코딩
+        byte[] decodedKey = Base64.getDecoder().decode(secretKey);
+        return Keys.hmacShaKeyFor(decodedKey);
+    }
+
+}
