+++ src/main/java/com/multi/travel/auth/jwt/domain/RefreshToken.java
@@ -1,0 +1,32 @@
+package com.multi.travel.auth.jwt.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name = "refresh_token")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class RefreshToken {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    @Column(nullable = false, columnDefinition = "TEXT")
+    private String refreshToken;
+
+    @Column(nullable = false)
+    private LocalDateTime expiredAt;
+
+    @Column(nullable = false)
+    private LocalDateTime issuedAt;
+}
