+++ src/main/java/com/multi/travel/auth/jwt/repository/RefreshTokenRepository.java
@@ -1,0 +1,18 @@
+package com.multi.travel.auth.jwt.repository;
+
+
+import com.multi.travel.auth.jwt.domain.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+
+    Optional<RefreshToken> findByEmail(String email);
+
+    @Modifying
+    @Transactional
+    void deleteByEmail(String email);
+}
