+++ src/main/java/com/multi/travel/auth/user/enums/UserStatus.java
@@ -1,0 +1,16 @@
+package com.multi.travel.auth.user.enums;
+
+/*
+ * Please explain the class!!!
+ *
+ * @filename    : UserStatus
+ * @author      : Choi MinHyeok
+ * @since       : 25. 12. 1. 월요일
+ */
+
+
+public enum UserStatus {
+    ACTIVE,       // 정상
+    INACTIVE,     // 휴면
+    DELETED,      // 탈퇴(소프트 삭제)
+}
