--- src/main/java/com/multi/travel/common/domain/BaseTimeEntity.java
@@ -1,11 +1,0 @@
-package com.multi.travel.common.domain;
-
-/**
- * Please explain the class!!!
- *
- * @author : rlagkdus
- * @filename : BaseTimeEntity
- * @since : 2025. 11. 8. 토요일
- */
-public class BaseTimeEntity {
-}
