--- src/main/java/com/multi/travel/common/exception/DuplicateUsernameException.java
@@ -1,7 +1,0 @@
-package com.multi.travel.common.exception;
-
-public class DuplicateUsernameException extends RuntimeException {
-    public DuplicateUsernameException(String message) {
-        super(message);
-    }
-}
