--- src/main/java/com/multi/travel/common/jwt/JwtAcessDeniedHandler.java
@@ -1,27 +1,0 @@
-package com.multi.travel.common.jwt;
-
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-@Slf4j
-public class JwtAcessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
-        log.warn("접근 거부: 사용자가 {} 경로에 접근할 권한이 없습니다. 403 Forbidden 반환.", request.getRequestURI());
-
-        response.setStatus(HttpServletResponse.SC_FORBIDDEN);
-        response.setContentType("application/json;charset=UTF-8");
-        response.getWriter().write("{\"error\": \"접근 권한이 없습니다.\"}");
-        response.getWriter().flush();
-    }
-}
