--- src/main/java/com/multi/travel/common/jwt/TokenProvider.java
@@ -1,147 +1,0 @@
-package com.multi.travel.common.jwt;
-
-
-import com.multi.travel.auth.dto.CustomUser;
-import com.multi.travel.common.exception.TokenException;
-import io.jsonwebtoken.*;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.stereotype.Component;
-
-import java.security.Key;
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Component
-@Slf4j
-public class TokenProvider {
-
-    private static final String AUTHORITIES_KEY = "auth";  // claim 에서 권한정보담을 키
-    private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 60;     // 1000 * 60 * 60 --> 60분
-    private static final long REFRESH_TOKEN_EXPIRE_TIME =1000L * 60 * 60 * 24; //1000L * 60 * 60 * 24 --> 1일
-
-    private final JwtProvider jwtProvider;  // JwtProvider 의존성 추가
-    private final Key SKEY;
-    private final String ISSUER;
-
-    //application.yml 에 정의해놓은 jwt.secret 값을 가져와서 JWT 를 만들 때 사용하는 암호화 키값을 생성
-    public TokenProvider(JwtProvider jwtProvider) {
-        this.jwtProvider = jwtProvider;
-        SKEY = jwtProvider.getSecretKey();
-        ISSUER = jwtProvider.getIssuer();
-        log.info("TokenProvider SKEY ---------------{}", SKEY);
-        log.info("TokenProvider ISSUER ---------------{}", ISSUER);
-    }
-
-    public String generateToken(String email, List<String> roles, String code) {
-
-        Claims claims = Jwts
-                .claims()
-                .setSubject(email);
-
-        // ✅ 두 토큰 모두에 auth 클레임 추가 (역할정보 유지)
-        claims.put(AUTHORITIES_KEY, String.join(",", roles));
-
-        long now = (new Date()).getTime();
-        Date tokenExpiresIn;
-        if (code.equals("A")) {  //액세스 토큰
-            tokenExpiresIn = new Date(now + ACCESS_TOKEN_EXPIRE_TIME);
-        } else {  //리프레시 토큰
-            tokenExpiresIn = new Date(now + REFRESH_TOKEN_EXPIRE_TIME);
-        }
-
-
-
-        return Jwts.builder()
-                .setIssuer(ISSUER) // 프로퍼티에 설정한 발행자
-                .setIssuedAt(new Date(now)) // 발생일 , 현재 시간
-                .setClaims(claims)                        // payload "auth": "ROLE_USER"
-                .setExpiration(tokenExpiresIn)       // payload "exp": 1516239022 (예시) // exp : Expiration Time. 토큰 만료 시각
-                .signWith(SKEY, SignatureAlgorithm.HS512)   // header "alg": "HS512"  // "alg": "서명 시 사용하는 알고리즘",
-                .compact();
-    }
-
-    public LocalDateTime getRefreshTokenExpiry() {
-        LocalDateTime now = LocalDateTime.now();
-
-        return now.plus(REFRESH_TOKEN_EXPIRE_TIME, ChronoUnit.MILLIS);
-    }
-
-    public boolean validateToken(String token) {
-        try {
-            log.info("[TokenProvider] 유효성 검증 중인 토큰: {}", token);
-            // 토큰을 비밀 키 와함께 복호화를 진행 해서 유효하지 않으면 false 반환, 유효하면 true 반환
-            Jwts.parserBuilder()
-                    .setSigningKey(SKEY)
-                    .build()
-                    .parseClaimsJws(token);
-
-            log.info("[TokenProvider] JWT 토큰이 유효합니다.");
-            return true;
-
-        } catch (io.jsonwebtoken.security.SecurityException | MalformedJwtException e) {
-            log.error("[TokenProvider] 잘못된 JWT 서명입니다. 토큰: {}", token, e);
-            throw new TokenException("잘못된 JWT 서명입니다.");
-
-        } catch (ExpiredJwtException e) {
-            log.warn("[TokenProvider] 만료된 JWT 토큰입니다. 토큰: {}, 만료 시각: {}", token, e.getClaims().getExpiration(), e);
-            throw new TokenException("만료된 JWT 토큰입니다.");
-
-        } catch (UnsupportedJwtException e) {
-            log.error("[TokenProvider] 지원되지 않는 JWT 토큰입니다. 토큰: {}", token, e);
-            throw new TokenException("지원되지 않는 JWT 토큰입니다.");
-
-        } catch (IllegalArgumentException e) {
-            log.error("[TokenProvider] JWT 토큰이 잘못되었습니다. 토큰: {}", token, e);
-            throw new TokenException("JWT 토큰이 잘못되었습니다.");
-        }
-    }
-
-    public Authentication getAuthentication(String jwt) {
-
-        Claims claims = parseClaims(jwt);
-        if (claims.get(AUTHORITIES_KEY) == null) {
-            throw new RuntimeException("권한정보가 없는 토큰입니다");
-
-        }
-        Collection<? extends GrantedAuthority> authorities =
-                Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
-                        .map(SimpleGrantedAuthority::new)
-                        .collect(Collectors.toList());
-
-        log.info("[TokenProvider] authorities : {}", authorities);
-
-        CustomUser customUser = new CustomUser();
-        customUser.setEmail(claims.getSubject());
-        customUser.setAuthorities(authorities);
-        return new UsernamePasswordAuthenticationToken(customUser, "", authorities);
-    }
-
-    public Claims parseClaims(String jwt) {
-
-        try {
-            return Jwts.parserBuilder()
-                    .setSigningKey(SKEY).build().parseClaimsJws(jwt).getBody();
-        }catch (ExpiredJwtException e){
-            return e.getClaims();
-        }
-    }
-
-    public String getEmail(String accessToken) {
-        try {
-            return Jwts.parserBuilder().setSigningKey(SKEY).build()
-                    .parseClaimsJws(accessToken).getBody().getSubject();
-        } catch (ExpiredJwtException e) {
-            return e.getClaims().getSubject(); //만료된 토큰에서도 email 가져오기
-        }
-    }
-
-}
