--- src/main/java/com/multi/travel/common/jwt/dto/TokenDto.java
@@ -1,17 +1,0 @@
-package com.multi.travel.common.jwt.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenDto {
-    @Builder.Default
-    private String grantType = "Bearer";
-    private String accessToken;
-    private String refreshToken;
-}
